Some refactoring. Remove unnecessary diffs.
