+++ 3-bull-jobs/finished-app/queues/GenreFetchingQueue.js
@@ -1,5 +1,5 @@
-const { BaseQueue } = require('./BaseQueue.js');
-const { QUEUES } = require('../constants.js');
+const { BaseQueue } = require('./BaseQueue');
+const { QUEUES } = require('../constants');
 
 const classParams = [
   QUEUES.GENRE_FETCHING_QUEUE.name,
