+++ 3-bull-jobs/finished-app/workers/index.js
@@ -1,7 +1,7 @@
-const { QUEUES } = require('../constants.js');
-const { FetchGenresWorker } = require('./FetchGenresWorker.js');
-const { FetchMoviesWorker } = require('./FetchMoviesWorker.js');
-const { SortMoviesWorker } = require('./SortMoviesWorker.js');
+const { QUEUES } = require('../constants');
+const { FetchGenresWorker } = require('./FetchGenresWorker');
+const { FetchMoviesWorker } = require('./FetchMoviesWorker');
+const { SortMoviesWorker } = require('./SortMoviesWorker');
 
 new FetchGenresWorker(QUEUES.GENRE_FETCHING_QUEUE.name).startWork();
 new FetchMoviesWorker(QUEUES.MOVIE_FETCHING_QUEUE.name).startWork();
